apps/api-server/requirements.txt
face-recognition
numpy
supabase
playwright
fastapi
uvicorn[standard]
pillow
python-multipart
requests
python-dotenv

# Opcional: Para embeddings de más dimensiones (256, 512, etc.)
deepface  # Requerido para upload_linkedin_mentors.py y add_deepface_embeddings.py

